+++ components/Button/react.stories.js
@@ -8,24 +8,25 @@ storiesOf('Button', module)
   .addParameters({ readme: { sidebar: Readme } })
   .addDecorator(withKnobs)
   .add('default', () => {
-    const variant = select(
-      'Variant',
-      {
-        Default: 'default',
-        Base: '',
-        Primary: 'primary',
-        Secondary: 'secondary',
-      },
-      'light',
-    )
+    const variant = select('Variant', {
+      Default: 'default',
+      Base: '',
+      Primary: 'primary',
+      Secondary: 'secondary',
+      Light: 'light',
+    })
     const size = select(
       'Size',
       {
         Small: 'small',
         Base: undefined,
+        Large: 'large',
       },
       '',
     )
+
+    const active = boolean('Active', false)
+
     const disabled = boolean('Disabled', false)
 
     const uppercase = boolean('Uppercase', false)
@@ -36,6 +37,7 @@ storiesOf('Button', module)
         size={size}
         uppercase={uppercase}
         disabled={disabled}
+        active={active}
       >
         Lorem Button
       </Button>