+++ components/LoadingModal/react.js
@@ -3,11 +3,9 @@ import { React as ModalBackdrop } from 'components/ModalBackdrop'
 import AnimatedMasoniteLogo from 'components/AnimatedMasoniteLogo'
 
 const LoadingModal = () => (
-
   <ModalBackdrop>
     <AnimatedMasoniteLogo />
   </ModalBackdrop>
-
 )
 
 export default LoadingModal