+++ components/OverflowMenu/react.stories.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
+import { withKnobs, select } from '@storybook/addon-knobs'
 import classNames from 'classnames/bind'
 import Button from 'components/Button/react'
 import OverflowMenu from './react'
@@ -10,22 +11,39 @@ const cx = classNames.bind(storyStyles)
 
 storiesOf('OverflowMenu', module)
   .addParameters({ readme: { sidebar: Readme } })
-  .add(
-    'default',
-    () => (
-      <OverflowMenu>
-        <p>saved on: 05/30/18<br />
+  .addDecorator(withKnobs)
+  .add('default', () => {
+    const variant = select(
+      'Variant',
+      {
+        Light: 'light',
+        Dark: 'dark',
+      },
+      'light',
+    )
+
+    return (
+      <OverflowMenu variant={variant}>
+        <p>
+          saved on: 05/30/18
+          <br />
           edited by: lorem ipsum
         </p>
         <span className={cx('overflow-menu__button-wrapper')}>
-          <Button variant="default" fullWidth uppercase>My First Button</Button>
+          <Button variant="default" fullWidth uppercase>
+            My First Button
+          </Button>
         </span>
         <span className={cx('overflow-menu__button-wrapper')}>
-          <Button variant="default" fullWidth uppercase>My Second Button</Button>
+          <Button variant="default" fullWidth uppercase>
+            My Second Button
+          </Button>
         </span>
         <span className={cx('overflow-menu__button-wrapper')}>
-          <Button variant="default" fullWidth uppercase>My Third Button</Button>
+          <Button variant="default" fullWidth uppercase>
+            My Third Button
+          </Button>
         </span>
       </OverflowMenu>
-    ),
-  )
+    )
+  })