+++ components/SidebarNav/react.js
@@ -29,11 +29,13 @@ const Submenu = ({
       'sidebar-nav-submenu--expanded': expanded,
     })}
   >
-    {items.map(({ label, href = '#', onClick }) => (
+    {items.map(({
+      label, href = '#', onClick, active = null,
+    }) => (
       <li key={label}>
         <a
           className={cx('menu-item-link', {
-            'menu-item-link--active': location && location.pathname === href,
+            'menu-item-link--active': active || (location && location.pathname === href),
           })}
           href={href}
           tabIndex="0"
@@ -88,6 +90,7 @@ const MainMenuItem = ({
         tabIndex="0"
         className={cx('menu-item-link')}
         href={href}
+        title={label}
         onClick={e => {
           if (onClick) {
             e.preventDefault()
@@ -165,7 +168,7 @@ export const SidebarNav = ({
         onTransitionEnd={() => setShowMainMenuLabels(showingMainMenuLabels)}
       >
         <ul className={cx('main-menu__top')}>
-          <li className={cx(['menu-item', 'menu-item--logo'])}>{logo}</li>
+          {logo && <li className={cx(['menu-item', 'menu-item--logo'])}>{logo}</li>}
           {menuItems.map((props, index) => (
             <MainMenuItem index={index} {...props} {...menuItemDependencies} />
           ))}
@@ -221,13 +224,19 @@ SidebarNav.defaultProps = {
   expandedCb: () => {},
 }
 
-export const SidebarNavLayout = props => {
+export const SidebarNavLayout = ({ layoutWrapperProps, contentWrapperProps, ...props }) => {
   const [expanded, setExpanded] = useState(false)
   const { children } = props
+
   return (
-    <div>
+    <div className={cx('sidebar-nav-wrapper')} {...layoutWrapperProps}>
       <SidebarNav {...props} expandedCb={expanded => setExpanded(expanded)} />
-      <div className={cx('content', expanded && 'content--main-menu-expanded')}>{children}</div>
+      <div
+        className={cx('content', expanded && 'content--main-menu-expanded')}
+        {...contentWrapperProps}
+      >
+        {children}
+      </div>
     </div>
   )
 }
@@ -236,4 +245,12 @@ SidebarNavLayout.propTypes = {
   children: PropTypes.node.isRequired,
 }
 
+SidebarNavLayout.defaultProps = {
+  layoutWrapperProps: null,
+  contentWrapperProps: null,
+}
+
 export default SidebarNavLayout
+
+const setInlineSvgTitle = (svg, title) =>
+  svg.replace(/<title>\w+<\/title>/, `<title>${title}</title>`)