+++ components/TextField/react.stories.js
@@ -15,50 +15,113 @@ storiesOf('TextField', module)
     'default',
     withState({ value: '' })(({ store }) => {
       const variant = select('Variant', { Light: 'light', Dark: 'dark' }, 'dark')
-      const align = select('Align', { Left: '', Right: 'right', Center: 'center' }, '')
+      const align = select('Align', { Left: 'left', Right: 'right', Center: 'center' }, 'left')
       const label = text('Label', 'Label')
+      const labelAlwaysAbove = boolean('Label always above', false)
+      const placeholder = text('Placeholder', 'Placeholder')
       const error = boolean('Error State', false)
       const helper = error ? 'Something is wrong.' : ''
       const showTrailingIcon = boolean('Show Trailing Icon', true)
+      const disabled = boolean('Disabled', false)
+      const readonly = boolean('Readonly', false)
+      const removeMargin = boolean('Remove margins', false)
       const trailingIcon = showTrailingIcon && mPhone
+
       return (
         <TextField
-          error={error}
-          helper={helper}
-          label={label}
-          placeholder="Placeholder"
+          {...{
+            align,
+            disabled,
+            error,
+            helper,
+            label,
+            labelAlwaysAbove,
+            placeholder,
+            readonly,
+            removeMargin,
+            trailingIcon,
+            variant,
+          }}
           onBlur={value => {}}
           onChange={value => store.set({ value })}
-          trailingIcon={trailingIcon}
           value={store.state.value}
-          variant={variant}
-          removeMargin={false}
-          align={align}
         />
       )
     }),
   )
   .add(
-    'center',
+    'stacked',
     withState({ value: '' })(({ store }) => {
       const variant = select('Variant', { Light: 'light', Dark: 'dark' }, 'dark')
+      const align = select('Align', { Left: 'left', Right: 'right', Center: 'center' }, 'left')
+      const label = text('Label', 'Label')
+      const labelAlwaysAbove = boolean('Label always above', false)
+      const placeholder = text('Placeholder', 'Placeholder')
       const error = boolean('Error State', false)
       const helper = error ? 'Something is wrong.' : ''
       const showTrailingIcon = boolean('Show Trailing Icon', true)
+      const disabled = boolean('Disabled', false)
+      const readonly = boolean('Readonly', false)
+      const removeMargin = boolean('Remove margins', false)
       const trailingIcon = showTrailingIcon && mPhone
+
       return (
-        <TextField
-          error={error}
-          helper={helper}
-          placeholder="Placeholder"
-          onBlur={value => console.log('onBlur called, value: ', value)}
-          onChange={value => store.set({ value })}
-          trailingIcon={trailingIcon}
-          value={store.state.value}
-          variant={variant}
-          removeMargin={false}
-          align="center"
-        />
+        <>
+          <TextField
+            {...{
+              align,
+              disabled,
+              error: true,
+              helper,
+              label,
+              labelAlwaysAbove,
+              placeholder,
+              readonly,
+              removeMargin,
+              trailingIcon,
+              variant,
+            }}
+            onBlur={value1 => {}}
+            onChange={value1 => store.set({ value1 })}
+            value={store.state.value1}
+          />
+          <TextField
+            {...{
+              align,
+              disabled,
+              error,
+              helper,
+              label,
+              labelAlwaysAbove,
+              placeholder,
+              readonly,
+              removeMargin,
+              trailingIcon,
+              variant,
+            }}
+            onBlur={value2 => {}}
+            onChange={value2 => store.set({ value2 })}
+            value={store.state.value2}
+          />
+          <TextField
+            {...{
+              align,
+              disabled,
+              error,
+              helper,
+              label,
+              labelAlwaysAbove,
+              placeholder,
+              readonly,
+              removeMargin,
+              trailingIcon,
+              variant,
+            }}
+            onBlur={value3 => {}}
+            onChange={value3 => store.set({ value3 })}
+            value={store.state.value3}
+          />
+        </>
       )
     }),
   )