+++ components/WizardModal/react.stories.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
+import { withKnobs, boolean } from '@storybook/addon-knobs'
 import { mGlobe } from '@masonite/svg-icons'
 import { React as ModalBackdrop } from 'components/ModalBackdrop'
 import { React as SmallTitle } from 'components/SmallTitle'
@@ -10,24 +11,28 @@ import Readme from './react.README.md'
 
 storiesOf('WizardModal', module)
   .addParameters({ readme: { sidebar: Readme } })
-  .add('example', () => (
-    <ModalBackdrop>
-      <Modal>
-        <Modal.Header
-          title="Hello World"
-          subtitle="Subtitle or Subheading"
-          icon={mGlobe}
-          onDismiss={() => { }}
-        />
-        <Modal.Body>
-          <SmallTitle>Small Title as Field Label</SmallTitle>
-          <TextField placeholder="Some Field" onChange={() => { }} value="" variant="light" />
-        </Modal.Body>
-        <Modal.Footer activePage={0} totalPages={4}>
-          <Button variant="secondary" uppercase>
-            Save
-          </Button>
-        </Modal.Footer>
-      </Modal>
-    </ModalBackdrop>
-  ))
+  .addDecorator(withKnobs)
+  .add('example', () => {
+    const show = boolean('Show', true)
+    return (
+      <ModalBackdrop show={show}>
+        <Modal>
+          <Modal.Header
+            title="Hello World"
+            subtitle="Subtitle or Subheading"
+            icon={mGlobe}
+            onDismiss={() => {}}
+          />
+          <Modal.Body>
+            <SmallTitle>Small Title as Field Label</SmallTitle>
+            <TextField placeholder="Some Field" onChange={() => {}} value="" variant="light" />
+          </Modal.Body>
+          <Modal.Footer activePage={0} totalPages={4}>
+            <Button variant="secondary" uppercase>
+              Save
+            </Button>
+          </Modal.Footer>
+        </Modal>
+      </ModalBackdrop>
+    )
+  })