+++ packages/svg-icons/build.js
@@ -13,7 +13,7 @@ try {
 const exportStatements = svgFiles
   .map(iconFilename => {
     const hyphenatedName = iconFilename.slice(0, -4)
-    const camelcaseName = hyphenatedName.replace(/-([a-z0-9])/g, g => (g.match(/[a-z]/) ? g[1].toUpperCase() : g[1]))
+    const camelcaseName = hyphenatedName.replace(/-([A-Za-z0-9])/g, g => (g.match(/[a-z]/) ? g[1].toUpperCase() : g[1]))
     const exportName = `${camelcaseName[0].toUpperCase()}${camelcaseName.slice(1)}`
 
     const inlineSvgExport = `export { default as m${exportName} } from './../../../vendor/micons/svg/SVG/${hyphenatedName}.svg'`